--- beauty/src/main/java/kr/co/beauty/dao/Order1DAO.java
@@ -1,30 +1,0 @@
-/* 
- * 김동근
- * CartVO
- * WishVO
- */
-package kr.co.beauty.dao;
-
-import java.util.List;
-
-import org.apache.ibatis.annotations.Mapper;
-import org.apache.ibatis.annotations.Param;
-import org.springframework.stereotype.Repository;
-
-import kr.co.beauty.vo.CartVO;
-
-@Mapper
-@Repository
-public interface Order1DAO {
-	
-	//cart
-	public List<CartVO> selectCartList(@Param("uid") String uid);
-	public void deleteSelectedCart(@Param("cartNo") int cartNo);
-	public void deleteAllCart(@Param("uid") String uid);
-	
-	//cart - change count
-	public int checkCountForUpdate(@Param("cartNo") int cartNo);
-	public int cartIncrease(@Param("cartNo") int cartNo);
-	public int cartDecrease(@Param("cartNo") int cartNo);
-
-}
