+++ beauty/src/main/resources/mappers/order.xml
@@ -2,8 +2,37 @@
 <!DOCTYPE mapper
   PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
   "https://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<!-- namespace = mapper파일의 아이디 값 -->
-<mapper namespace="kr.co.beauty.dao.Order2DAO">
+<!-- 김동근 -->
+<mapper namespace="kr.co.beauty.dao.OrderDAO">
+	
+	<!-- 카트 -->
+	<select id="selectCartList" resultType="CartVO">
+		SELECT * FROM `cart` WHERE `uid`=#{uid};
+	</select>
+	
+	<delete id="deleteSelectedCart">
+		DELETE FROM `cart` WHERE `cartNo`=#{cartNo};
+	</delete>
+	<delete id="deleteAllCart">
+		DELETE FROM `cart` WHERE `uid`=#{uid};
+	</delete>
+	
+	<select id="checkCountForUpdate" resultType="int">
+		SELECT `count` FROM `cart`
+		WHERE `cartNo` = #{cartNo};
+	</select>
+	<update id="cartIncrease">
+		UPDATE `cart` SET
+		count = count + 1
+		WHERE `cartNo` = #{cartNo};
+	</update>
+	<update id="cartDecrease">
+		UPDATE `cart` SET
+		count = count - 1
+		WHERE `cartNo` = #{cartNo};
+	</update>
+
+	<!-- 박진휘 -->	
 	<select id="selectOrdercomplete" resultType="OrdercompleteVO">
 		SELECT * FROM `ordercomplete` WHERE `ordNo`=#{ordNo};
 	</select>
@@ -42,4 +71,5 @@
 		#{orderer},#{orderHp},#{orderZip},#{orderAddr1},#{orderAddr2},#{orderEmail},
 		#{recipName},#{recipHp},#{recipZip},#{recipAddr1},#{recipAddr2},#{message},#{payment},NOW());
 	</insert>
+
 </mapper>
