+++ beauty/src/main/resources/static/css/order/cart.css
@@ -37,9 +37,10 @@ main div.cartFrame > .nav > p.current span {
     background: #81cb49;
 }
 
-/*마이페이지 - 콘텐츠 - 위시리스트*/
+/*마이페이지 - 콘텐츠 - 카트*/
 main div.cartFrame {
     width: 1200px;
+    min-height: 350px;
     margin: 0 auto;
 }
 
@@ -408,9 +409,11 @@ main div.wishListFrame > .wishList > .item > .itemInfo > span {
 	margin: 30px 0;	
 }
 .cartFrame > article.cartList > div.emptyCart {
+	height:150px;
 	color: #555555;
 	font-weight: bold;
 	font-size: 20px;
 	text-align: center;
+	line-height: 150px;
 	margin: 30px 0;	
 }
