+++ beauty/src/main/resources/static/js/order/cart.js
@@ -89,6 +89,15 @@ $(function(){
 	
 
 	/* tableBtns */
+	/* 주문하기 */
+	$('div.cartFrame > article.cartList > table > tbody td > a.btnTableOrder').click(function(e){
+		e.preventDefault();
+		let chkList = [];
+		chkList.push(parseInt($(this).parent().parent().children('td:first-child').children('input').val()));
+		redirectGet('/Beauty/order/orderform/type1', chkList);
+	});
+
+	
 	/* 관심상품등록 */
 	$('div.cartFrame > article.cartList > table > tbody td > a.btnTableAddWish').click(function(e){
 		e.preventDefault();
@@ -153,6 +162,11 @@ $(function(){
 	/* 카트 선택 삭제 */
 	$('#btnDeleteSelectedCart').click(function(e){
 		e.preventDefault();
+		if ($("input:checkbox[name='chkCart']").is(":checked")==false) {
+			/* 선택 안 한 경우 */
+			alert('삭제하실 상품을 선택해주세요.');
+			return;
+		}
 		let chkList = [];
 		$('input[name=chkCart]').each(function(){
 			if($(this).is(":checked")){
@@ -187,7 +201,112 @@ $(function(){
 			}
 		});
 	});
+	
+	/* cartBtns Order */
+	/* 선택주문하기 */
+	$('#btnOrderSelect').click(function(e){
+		e.preventDefault();
+		if ($("input:checkbox[name='chkCart']").is(":checked")==false) {
+			/* 선택 안 한 경우 */
+			alert('주문하실 상품을 선택해주세요.');
+			return;
+		}
+		let chkList = [];
+		$('input[name=chkCart]').each(function(){
+			if($(this).is(":checked")){
+				chkList.push(parseInt($(this).val()));
+			}
+		});
+		redirectGet('/Beauty/order/orderform/type1', chkList);
+	});
+	/* 전체주문하기 */
+	$('#btnOrderAll').click(function(e){
+		e.preventDefault();
+		let chkList = [];
+		$('input[name=chkCart]').each(function(){
+			chkList.push(parseInt($(this).val()));
+		});
+		redirectGet('/Beauty/order/orderform/type1', chkList);
+	});
+	
+	
+	
+	
+	
+	
+		/* No-member  */
+	/* tableBtns */
+	/* No-member 주문하기 */
+	$('div.cartFrame > article.cartList > table > tbody td > a.btnTableOrderNon').click(function(e){
+		e.preventDefault();
+		let chkList = [];
+		chkList.push(parseInt($(this).parent().parent().children('td:first-child').children('input').val()));
+		redirectGet('/Beauty/order/orderform/type2', chkList);
+	});
+	
+	/* No-member 관심상품등록 */
+	$('div.cartFrame > article.cartList > table > tbody td > a.btnTableAddWishNon').click(function(e){
+		e.preventDefault();
+		alert('로그인이 필요합니다');
+		//로그인창이동 or 그대로
+		
+	});
 
+	/* cartBtns */
+	/* No-member 카트비우기 */
+	$('#btnDeleteAllCartNon').click(function(e){
+		e.preventDefault();
+		if(confirm('장바구니를 비우시겠습니까?')){
+			$.ajax({
+				url:'/Beauty/order/deleteAllCartNon',
+				type:'POST',
+				data:{},
+				dataType:'json',
+				success:function(data){
+					if(data == 1){
+						emptyTable();
+					}else{
+					}
+				}
+			});
+		}else{
+			return;
+		}
+	});
+	
+	/* cartBtns Order */
+	/* No-member 선택주문하기 */
+	$('#btnOrderSelectNon').click(function(e){
+		e.preventDefault();
+		if ($("input:checkbox[name='chkCart']").is(":checked")==false) {
+			/* 선택 안 한 경우 */
+			alert('주문하실 상품을 선택해주세요.');
+			return;
+		}
+		let chkList = [];
+		$('input[name=chkCart]').each(function(){
+			if($(this).is(":checked")){
+				chkList.push(parseInt($(this).val()));
+			}
+		});
+		redirectGet('/Beauty/order/orderform/type1', chkList);
+	});
+	/* No-member 전체주문하기 */
+	$('#btnOrderAllNon').click(function(e){
+		e.preventDefault();
+		let chkList = [];
+		$('input[name=chkCart]').each(function(){
+			chkList.push(parseInt($(this).val()));
+		});
+		redirectGet('/Beauty/order/orderform/type1', chkList);
+	});
+	
+	
+	
+	
+	
+	
+	
 	/* 빈 카트 */
 	function emptyTable(){
 		//카트 비었으면
@@ -197,9 +316,6 @@ $(function(){
 		$('#headCount').text("0");
 		$('div.cartFrame > article.cartList').append('<div class="emptyCart">장바구니가 비어 있습니다.</div>');
 	}
-
-
-
 
 	/* 선택 구분 함수 */	
 	function sumChecked(){
@@ -257,6 +373,25 @@ $(function(){
 		$('#totalDelivery').text(delivery);
 		$('#totalDisPrice').text(disPrice.toLocaleString()+'원');
 		$('#totalTotalPrice').text(total.toLocaleString()+'원');
+	}
+	
+	/* form 전송 */
+	function redirectGet(location, args){
+		let form = $('<form></form>');
+        form.attr("method", "get");
+        form.attr("action", location);
+        
+        $.each(args, function (index, value) {
+            let field = $('<input></input>');
+            
+            field.attr('type', 'hidden');
+            field.attr("name", 'cartNo');
+            field.attr("value", value);
+            
+            form.append(field);
+        });
+        
+        $(form).appendTo('body').submit();
 	}
 
 });
